backend/requirements.txt
# GPS Tracking System - Backend Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# WebSocket / Real-time
python-socketio==5.11.0
websockets==12.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client (for simulator)
httpx==0.26.0
aiohttp==3.9.3
shapely==2.0.3

# Utilities
python-dateutil==2.8.2
python-jose==3.3.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
